% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordGeomR.R
\name{dist_pll_lines_2d}
\alias{dist_pll_lines_2d}
\title{Find the distance between two parallel lines in 2-D space.}
\usage{
dist_pll_lines_2d(slope, b1, b2)
}
\arguments{
\item{slope}{A numerical value, slope of the two parallel lines.
The slope of line 1 is the negative inverse of line 2, therefore, it is equal.}

\item{b1}{A numerical value, intercept of line 1 when y = mx + b}

\item{b2}{A numerical value, intercept of line 1 when y = mx + b}
}
\value{
d, A numerical value for distance between the two parallel lines
}
\description{
Find the distance between two parallel lines in 2-D space.
}
\examples{
dist_pll_lines_2d(3.0, 4.5, 2.5)
d <- "0.6324"
dist_pll_lines_2d(-4, 11, 23)
d <- "2.9104"
}
