% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordGeomR.R
\name{get_distance}
\alias{get_distance}
\title{Calculates the distance between two n dimensional vectors.}
\usage{
get_distance(x1, x2, metric = "Euclidean", p = NULL)
}
\arguments{
\item{x1}{The first n dimensional vector.}

\item{x2}{The second n dimensional vector.}

\item{metric}{The distance metric, must be one of "Euclidean", "Manhattan", Chebyshev, or "Minkowski"}

\item{p}{The order of Minkowski distance to calculate.  Only required if \code{metric} is set to "Minkowski".}
}
\value{
The relevant distance between the two vectors.
}
\description{
Calculates the distance between two n dimensional vectors.
}
\examples{
x1 <- c(1, 2, 3, 4)
x2 <- c(5, 6, 7, 8)
get_distance(x1, x2, metric = "Euclidean")
8
get_distance(x1, x2, metric = "Manhattan")
16
get_distance(x1, x2, metric = "Chebyshev")
4
get_distance(x1, x2, metric = "Minkowski", 3)
6.3496042078728
}
