% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordGeomR.R
\name{is_orthogonal}
\alias{is_orthogonal}
\title{Determines whether two infinite lines are perpendicular in n-dimensional space.}
\usage{
is_orthogonal(m1, m2)
}
\arguments{
\item{m1}{A numeric vector corresponds to a n-dimensional vector ⟨mx1, my1, mz1, ...⟩ describing the direction vector of line 1.}

\item{m2}{A numeric vector corresponds to a n-dimensional vector ⟨mx1, my1, mz1, ...⟩ describing the direction vector of line 1. Demensions of line 1 and line 2 must be equal.}
}
\value{
TRUE if there is orthogonal, FALSE if not.
}
\description{
Determines whether two infinite lines are perpendicular in n-dimensional space.
}
\examples{
m1 <- c(1, 0, 0)
m2 <- c(0, 1, 0)
is_orthogonal(m1, m2)
}
